--- purchase_order_line_sequence/readme/USAGE.md
@@ -1,3 +1,0 @@
-This module depends on the following module(s) :  
-- stock_picking_line_sequence
-  (<https://github.com/OCA/stock-logistics-workflow/>)
