+++ purchase_order_line_sequence/tests/__init__.py
@@ -1,3 +1,4 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
+from . import common
 from . import test_po_lines_sequence
